chore: Maintenance - Migrate from CodeClimate to Qlty - Configure Qlty - Migrate from CodeCov to Qlty - Fix findings - Stop using PR labeler - Eliminate repo name literals from workflows
{patch}

+++ src/main/java/com/github/nagyesta/cacheonly/core/CacheRefreshStrategy.java
@@ -50,7 +50,7 @@ public enum CacheRefreshStrategy {
                 final List<I> fetchExtra = new ArrayList<>(idsFoundInCache);
                 Collections.shuffle(fetchExtra);
                 fetchExtra.stream()
-                        .limit(maxPartitionSize - mustBeInLastPartition)
+                        .limit((long) maxPartitionSize - mustBeInLastPartition)
                         .forEach(result::add);
             }
             return result;
